\name{ggExtra-package}
\alias{ggExtra-package}
\alias{ggExtra}
\docType{package}
\title{
ggExtra
}
\description{
A packages plots ggplot2 style figures but with additional features and functions.
}
\details{
\tabular{ll}{
Package: \tab ggExtra\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-19\cr
License: \tab GPL \cr
}
}
\author{
Tony Huang

huangbaochenwo@live.com
}
\references{
Wickham, H. (2009). ggplot2: elegant graphics for data analysis (Use R): Springer.
Chang, W. (2012). R graphics cookbook: " O'Reilly Media, Inc.".
}
\keyword{ package }
\seealso{

}
