% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggExtra.R
\docType{package}
\name{ggExtra}
\alias{ggExtra}
\alias{ggExtra-package}
\title{ggExtra}
\description{
Collection of functions and layers to enhance \code{ggplot2}. The main
function is \code{\link[ggExtra]{ggMarginal}}, which can be used to add
marginal histograms/boxplots/density plots to ggplot2 plots.
}
\details{
View a \href{http://daattali.com/shiny/ggExtra-ggMarginal-demo/}{demo Shiny app}
or see the full \href{https://github.com/daattali/ggExtra}{README} on GitHub.
}

