% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMarginalGadget.R
\name{ggMarginalGadget}
\alias{ggMarginalGadget}
\title{ggMarginal gadget}
\usage{
ggMarginalGadget(plot)
}
\arguments{
\item{plot}{A ggplot2 scatterplot}
}
\value{
An object of class \code{ggExtraPlot}. This object can be printed to
show the marginal plots or saved using any of the typical image-saving functions
}
\description{
This gadget and addin allow you to select a ggplot2 plot and interactively
use \code{ggMarginal} to build marginal plots on top of your scatterplot.
}
\note{
To use the RStudio addin, highlight the code for a plot in RStudio and
select \emph{ggplot2 Marginal Plots} from the RStudio \emph{Addins} menu. This will
embed the marginal plots code into your script. Alternatively, you can call
\code{ggMarginalGadget()} with a ggplot2 plot, and the gadget will return
a plot object.
}
\examples{
if (interactive()) {
  myplot <- ggplot2::ggplot(mtcars, ggplot2::aes(wt, mpg)) + ggplot2::geom_point()
  plot2 <- ggMarginalGadget(myplot)
}
}
