\name{ggplotExtra-package}
\alias{ggplotExtra-package}
\alias{ggplotExtra}
\docType{package}
\title{
ggExtra
}
\description{
A packages plots ggplot2 style figures but with additional features and functions
}
\details{
\tabular{ll}{
Package: \tab ggplotExtra\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-19\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Tony Huang

huangbaochenwo@live.com
}
\references{
Wickham, H. (2009). ggplot2: elegant graphics for data analysis (Use R): Springer.
Chang, W. (2012). R graphics cookbook: " O'Reilly Media, Inc.".
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
