% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMarginal.R
\name{print.ggExtraPlot}
\alias{print.ggExtraPlot}
\title{Print a ggExtraPlot object}
\usage{
\method{print}{ggExtraPlot}(x, newpage = grDevices::dev.interactive(), ...)
}
\arguments{
\item{x}{ggExtraPlot object.}

\item{newpage}{Should a new page (i.e., an empty page) be drawn before the
ggExtraPlot is drawn?}

\item{...}{ignored}
}
\description{
\code{ggExtraPlot} objects are created from \code{ggMarginal}. This is the S3
generic print method to print the result of the scatterplot with its marginal
plots.
}
\seealso{
\code{\link{ggMarginal}}
}
\keyword{internal}
