% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run ggExtra example [DEPRECATED]}
\usage{
runExample()
}
\description{
Launch a Shiny app that shows a demo of what can be done with
\code{ggExtra::ggMarginal}.
}
\details{
This example is also
\href{https://daattali.com/shiny/ggExtra-ggMarginal-demo/}{available online}.\cr\cr
\strong{Deprecation Notice:} This function is no longer required since Shiny version
1.8.1 (March 2024). This function will be removed in a future release of \{ggExtra\}.
You can use \code{shiny::runExample("ggMarginal", package = "ggExtra")} instead of
\code{ggExtra::runExample()}.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  runExample()
}
}
