% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggExtra-package.R
\docType{package}
\name{ggExtra-package}
\alias{ggExtra}
\alias{ggExtra-package}
\title{ggExtra: Add Marginal Histograms to 'ggplot2', and More 'ggplot2' Enhancements}
\description{
Collection of functions and layers to enhance 'ggplot2'. The flagship function is 'ggMarginal()', which can be used to add marginal histograms/boxplots/density plots to 'ggplot2' scatterplots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daattali/ggExtra}
  \item \url{https://daattali.com/shiny/ggExtra-ggMarginal-demo/}
  \item Report bugs at \url{https://github.com/daattali/ggExtra/issues}
}

}
\author{
\strong{Maintainer}: Dean Attali \email{daattali@gmail.com}

Authors:
\itemize{
  \item Christopher Baker \email{chriscrewbaker@gmail.com}
}

}
\keyword{internal}
