% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggMarginal.R
\name{ggMarginal}
\alias{ggMarginal}
\title{Add marginal density/histogram to ggplot2 scatterplots}
\usage{
ggMarginal(p, data, x, y, type = "density", margins = "both", size = 5,
  plot = TRUE)
}
\arguments{
\item{x}{Whether to apply th layer to the X axis.}
}
\description{
Create a ggplot2 scatterplot with marginal density plots/histograms or add
the marginal plots to an existing scatterplot.
}

