% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Run ggExtra example}
\usage{
runExample()
}
\description{
Launch a Shiny app that shows a demo of what can be done with
\code{ggExtra::ggMarginal}.
}
\details{
This example is also
\href{https://daattali.com/shiny/ggExtra-ggMarginal-demo/}{available online}.
}
\examples{
## Only run this example in interactive R sessions
if (interactive()) {
  runExample()
}
}
